修改adlist len数据类型

+++ structure/adlist_test.go
@@ -82,7 +82,7 @@ func TestList_Index(t *testing.T) {
 
 	index := -1000
 
-	fmt.Println(l.Index(int32(index)))
+	fmt.Println(l.Index(index))
 }
 
 func TestList_RotateTailToHead(t *testing.T) {
